install: Fix nesting symlinks

+++ install.sh
@@ -3,20 +3,46 @@
 # Installs configuration files, packages, and plugins
 set -e
 
+function dir() {
+  local target_dir="${1}"
+
+  if [ -d ${target_dir} ]; then
+    echo "${target_dir} already exists. Skipping."
+  elif [ -e ${target_dir} ]; then
+    echo "${target_dir} already exists, but is not a directory. Skipping."
+  else
+    mkdir --parents --verbose ${target_dir}
+  fi
+}
+
 function link() {
   local target_file="${1}"
   local link_name="${2}"
 
-  # -s - Symbolic link
-  # -f - If the link already exists, overwrite it
-  # -v - Verbose output
-  ln -sfv ${target_file} ${link_name}
+  if [ -L ${link_name} ]; then
+    echo "${link_name} already exists. Skipping."
+  elif [ -e ${link_name} ]; then
+    echo "${link_name} already exists, but is not a link. Skipping."
+  else
+    # -s - Symbolic link
+    # -f - If the link already exists, overwrite it
+    # -v - Verbose output
+    ln -sfv ${target_file} ${link_name}
+  fi
 }
 
 # Set vars
 BASEDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
+# Directories
+echo "###########################"
+echo "# Creating Directories"
+echo "###########################"
+dir "${HOME}/.config"
+
+
 # Symlinks
+echo
 echo "###########################"
 echo "# Creating Symlinks"
 echo "###########################"
@@ -28,10 +54,10 @@ link "${BASEDIR}/bash/rc"             "${HOME}/.bashrc"
 link "${BASEDIR}/git/gitignore"       "${HOME}/.gitignore"
 link "${BASEDIR}/git/git-sh-prompt"   "${HOME}/.git-prompt.sh"
 link "${BASEDIR}/intellij/ideavimrc"  "${HOME}/.ideavimrc"
-link "${BASEDIR}/i3/"                 "${HOME}/.config/i3"
-link "${BASEDIR}/neovim/"             "${HOME}/.config/nvim"
-link "${BASEDIR}/neovim/"             "${HOME}/.vim"
-link "${BASEDIR}/ranger/"             "${HOME}/.config/ranger"
+link "${BASEDIR}/i3"                  "${HOME}/.config/i3"
+link "${BASEDIR}/neovim"              "${HOME}/.config/nvim"
+link "${BASEDIR}/neovim"              "${HOME}/.vim"
+link "${BASEDIR}/ranger"              "${HOME}/.config/ranger"
 
 # Packages
 echo
